server/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1
pydantic[email]==2.10.5
pydantic-settings==2.7.1
python-jose[cryptography]==3.3.0
httpx==0.28.1
websockets==14.1
sse-starlette==2.2.1
resend==2.0.0
